pyproject.toml
[tool.poetry]
name = "resume-test"
version = "0.1.0"
description = ""
authors = ["Yuriy S. <trupakenator@gmail.com>"]
readme = "README.md"
packages = [{include = "resume_test"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"
pytest-factoryboy = "^2.5.1"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
